feat: add --dry-run option to build-colors script, add pro chem ice dye images

+++ scripts/build-colors.js
@@ -2,6 +2,7 @@ import * as prettier from "prettier";
 import * as fs from "fs";
 import * as path from "path";
 import { sortFn as byColor } from "color-sorter";
+import chalk from "chalk";
 
 import buildDharmaColors from "./build-colors-dharma.js";
 import buildProChemColors from "./build-colors-pro-chem.js";
@@ -11,13 +12,14 @@ const MANUFACTURER_BUILDERS = {
   ["pro-chem"]: buildProChemColors,
 };
 
-const manufacturerArgs = process.argv.slice(2);
+const args = process.argv.slice(2);
+const manufacturerArgs = args.filter((arg) => !/^--/.test(arg));
 
 const getFilteredBuilders = () => {
   return manufacturerArgs.reduce((acc, m) => {
     const builder = MANUFACTURER_BUILDERS[m];
     if (!builder) {
-      console.log(`No builder for manufacturer '${m}'`);
+      console.log(chalk.yellow(`No builder for manufacturer '${m}'`));
       return acc;
     }
 
@@ -31,8 +33,10 @@ const colorBuilders =
     : Object.values(MANUFACTURER_BUILDERS);
 
 if (colorBuilders.length === 0) {
-  console.warn(
-    `No builders found for manufacturers [${manufacturerArgs.join(",")}]. Exiting.`,
+  console.log(
+    chalk.red(
+      `No builders found for manufacturers [${manufacturerArgs.join(",")}]. Exiting.`,
+    ),
   );
   process.exit(0);
 }
@@ -48,10 +52,17 @@ const output = await prettier.format(JSON.stringify(manufacturers), {
   parser: "json-stringify",
 });
 
-fs.writeFile(
-  path.join(import.meta.dirname, "../public/colors-compiled.json"),
-  output,
-  (error) => {
-    if (error) console.error(error);
-  },
-);
+const isDryRun = args.includes("--dry-run");
+
+if (isDryRun) {
+  console.log("DRY RUN: colors-compiled.json");
+  console.log(output);
+} else {
+  fs.writeFile(
+    path.join(import.meta.dirname, "../public/colors-compiled.json"),
+    output,
+    (error) => {
+      if (error) console.error(error);
+    },
+  );
+}
